changed integer fields to floats, after observing fractional ele in the wild.

+++ gpx.go
@@ -81,10 +81,10 @@ type BoundsType struct {
 type WptType struct {
 	Lat           float64     `xml:"lat,attr"`
 	Lon           float64     `xml:"lon,attr"`
-	Ele           *int        `xml:"ele,omitempty"`           // Elevation (in meters) of the point.
+	Ele           *float64    `xml:"ele,omitempty"`           // Elevation (in meters) of the point.
 	Time          *time.Time  `xml:"time,omitempty"`          // Creation/modification timestamp for element. Date and time in are in Univeral Coordinated Time (UTC), not local time! Conforms to ISO 8601 specification for date/time representation. Fractional seconds are allowed for millisecond timing in tracklogs.
 	Magvar        *float64    `xml:"magvar,omitempty"`        // Magnetic variation (in degrees) at the point
-	Geoidheight   *int        `xml:"geoidheight,omitempty"`   // Height (in meters) of geoid (mean sea level) above WGS84 earth ellipsoid.  As defined in NMEA GGA message.
+	Geoidheight   *float64    `xml:"geoidheight,omitempty"`   // Height (in meters) of geoid (mean sea level) above WGS84 earth ellipsoid.  As defined in NMEA GGA message.
 	Name          *string     `xml:"name,omitempty"`          // The GPS name of the waypoint. This field will be transferred to and from the GPS. GPX does not place restrictions on the length of this field or the characters contained in it. It is up to the receiving application to validate the field before sending it to the GPS.
 	Cmt           *string     `xml:"cmt,omitempty"`           // GPS waypoint comment. Sent to GPS as comment.
 	Desc          *string     `xml:"desc,omitempty"`          // A text description of the element. Holds additional information about the element intended for the user, not the GPS.
@@ -93,11 +93,11 @@ type WptType struct {
 	Sym           *string     `xml:"sym,omitempty"`           // Text of GPS symbol name. For interchange with other programs, use the exact spelling of the symbol as displayed on the GPS.  If the GPS abbreviates words, spell them out.
 	Type          *string     `xml:"type,omitempty"`          // Type (classification) of the waypoint.
 	Fix           *string     `xml:"fix,omitempty"`           // Type of GPX fix. ('none'|'2d'|'3d'|'dgps'|'pps')
-	Sat           *int        `xml:"sat,omitempty"`           // Number of satellites used to calculate the GPX fix.
-	Hdop          *int        `xml:"hdop,omitempty"`          // Horizontal dilution of precision.
-	Vdop          *int        `xml:"vdop,omitempty"`          // Vertical dilution of precision.
-	Pdop          *int        `xml:"pdop,omitempty"`          // Position dilution of precision.
-	Ageofdgpsdata *int        `xml:"ageofdgpsdata,omitempty"` // Number of seconds since last DGPS update.
+	Sat           *float64    `xml:"sat,omitempty"`           // Number of satellites used to calculate the GPX fix.
+	Hdop          *float64    `xml:"hdop,omitempty"`          // Horizontal dilution of precision.
+	Vdop          *float64    `xml:"vdop,omitempty"`          // Vertical dilution of precision.
+	Pdop          *float64    `xml:"pdop,omitempty"`          // Position dilution of precision.
+	Ageofdgpsdata *float64    `xml:"ageofdgpsdata,omitempty"` // Number of seconds since last DGPS update.
 	Dgpsid        *string     `xml:"dgpsid,omitempty"`        // ID of DGPS station used in differential correction.
 }
